✨ Restored old persons page, new one as Administration
